+++ src/main/java/ctoutweb/lalamiam/model/dto/AddProductDto.java
@@ -1,12 +1,9 @@
 package ctoutweb.lalamiam.model.dto;
-
-import java.math.BigInteger;
-
 public record AddProductDto(
         String name,
         Double price,
         String description,
         Integer preparationTime,
         String photo,
-        BigInteger storeId) {
+        Long storeId) {
 }
