+++ src/main/java/ctoutweb/lalamiam/model/dto/AddStoreDto.java
@@ -2,11 +2,10 @@ package ctoutweb.lalamiam.model.dto;
 
 import ctoutweb.lalamiam.model.WeeklyStoreSchedule;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public record AddStoreDto(
-        BigInteger proId,
+        Long proId,
         String name,
         String adress,
         String city,
