+++ src/main/java/ctoutweb/lalamiam/model/dto/CommandInformationDto.java
@@ -2,14 +2,13 @@ package ctoutweb.lalamiam.model.dto;
 
 import ctoutweb.lalamiam.model.ProductWithQuantity;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public record CommandInformationDto(
-        BigInteger storeId,
-        BigInteger commandId,
+        Long storeId,
+        Long commandId,
         LocalDate commandDate,
         LocalDateTime consultationDate,
         List<ProductWithQuantity> productsSelected) {}
