+++ src/main/java/ctoutweb/lalamiam/model/dto/PersitCommandDto.java
@@ -2,7 +2,6 @@ package ctoutweb.lalamiam.model.dto;
 
 import ctoutweb.lalamiam.model.ProductWithQuantity;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,8 +10,8 @@ import java.util.List;
  * Données nécessaires pour enregistrer une commande
  */
 public record PersitCommandDto(
-        BigInteger storeId,
-        BigInteger commandId,
+        Long storeId,
+        Long commandId,
         LocalDate commandDate,
         LocalDateTime consultationDate,
         List<ProductWithQuantity> selectProducts,
